requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
sqlalchemy>=1.4.0
pyodbc>=4.0.0
python-dotenv>=1.0.0
